Move existing routes into new files
