+++ src/routes/legacyRoutes.ts
@@ -1,9 +1,42 @@
-import { asArray, asBoolean, asObject, asString, asValue } from 'cleaners'
+import {
+  asArray,
+  asBoolean,
+  asNumber,
+  asObject,
+  asString,
+  asValue
+} from 'cleaners'
 import { Serverlet } from 'serverlet'
 
-import { User } from '../../models/User'
-import { ApiRequest } from '../../types/requestTypes'
-import { jsonResponse } from '../../types/responseTypes'
+import { Device } from '../models/Device'
+import { User } from '../models/User'
+import { ApiRequest } from '../types/requestTypes'
+import { jsonResponse } from '../types/responseTypes'
+
+/**
+ * The GUI names this `registerDevice`, and calls it at boot.
+ *
+ * POST /v1/device?deviceId=...
+ * Request body: asRegisterDeviceRequest
+ * Response body: unused
+ */
+export const registerDeviceV1Route: Serverlet<ApiRequest> = async request => {
+  const { json, log, query } = request
+  const { deviceId } = asRegisterDeviceQuery(query)
+  const clean = asRegisterDeviceRequest(json)
+
+  let device = await Device.fetch(deviceId)
+  if (device) {
+    await device.save(clean as any)
+    log('Device updated.')
+  } else {
+    device = new Device(clean as any, deviceId)
+    await device.save()
+    log(`Device registered.`)
+  }
+
+  return jsonResponse(device)
+}
 
 /**
  * The GUI names this `getNotificationState`,
@@ -166,4 +199,17 @@ const asEnableCurrencyBody = asObject({
 
 const asToggleStateBody = asObject({
   enabled: asBoolean
+})
+
+const asRegisterDeviceQuery = asObject({
+  deviceId: asString
+})
+
+const asRegisterDeviceRequest = asObject({
+  appId: asString,
+  tokenId: asString, // Firebase device token
+  deviceDescription: asString,
+  osType: asString,
+  edgeVersion: asString,
+  edgeBuildNumber: asNumber
 })
